Another possible improvement for White Dots
Even with all the fixes implemented today for White Dots, I got some in the "enzo" car demo scene :-(

Normally we use the flag -ffast-math to improve speed but it causes the isnan() and isinf() functions not to work.

As we want the maximum possible speed, we don't want to remove the flag -ffast-math. However, perhaps we could re-enable only the capability of isnan() and isinf() functions by using the flag "-fno-finite-math-only" in addition to "-ffast-math"?

So, instead of  "-O3 -ffast-math", we would have
"-O3 -ffast-math -fno-finite-math-only"

Please, try this possible change to see if/how it impacts the render speeds. If there is only a minimal impact in speed, I think it's worth to use "-fno-finite-math-only", because it would solve almost all possible cases of NaN, Inf, etc for good!

I hope this helps! Best regards.