Fix for negated color in negated Musgrave/Voronoi textures
When the Musgrave texture is mapped to the material color and the texture is inverted, negative values are passed to the color causing a negation in the color as well. That also happens in Voronoi textures when they are negated and have a high intensity scale. I've truncated any negative values from the texture before they are used for color generation. This loses negative values for color, but I think this is what should be expected acording to the bug http://www.yafaray.org/node/187 ?

 Changes to be committed:
	modified:   src/textures/layernode.cc