Possible speed up change in the NaN fix?
Thinking this again, perhaps we can make a small speed up change. Because we are using lengths these are (in theory) only positive values so we don't need to check that the asin argument is less than -1.0?  In my tests, checking only for > 1.0 is enough to remove the black NaN areas.

Not sure how much speed would we gain from removing the check for -1.0?

If you think this is a good idea, please merge. If you think the speed change will be very small and you want to keep the fix as it is (checking both +1.0 and -1.0) then please close this pull request.