Automatic Shadow Bias and Min Ray Dist calculation
This is a very experimental change. Until now the Shadow Bias and Min Ray Dists were constant and independent of the scene size. Due to numeric precission and fast calculations limitations it tended to cause black artifacts in scenes with long triangles and/or big objects.

With this change, the Shadow Bias and Min Ray Dist calculations will be automatically calculated and showed in the YafaRay console log as in the following example:

INFO: Scene: total scene dimensions: X=994.484, Y=994.484, Z=1001.61, volume=9.90586e+08, calculated Shadow Bias=0.500803, calculated Ray Min Dist=0.250401

This change needs to be extensively tested to ensure it does not introduce new issues as it affects the fundamental calculations for shadows, photon mapping and other integrators. In any case, I hope it helps.

 Changes to be committed:
	modified:   include/core_api/scene.h
	modified:   include/yafraycore/kdtree.h
	modified:   include/yafraycore/ray_kdtree.h
	modified:   src/integrators/SingleScatterIntegrator.cc
	modified:   src/integrators/bidirpath.cc
	modified:   src/integrators/pathtracer.cc
	modified:   src/integrators/photonintegr.cc
	modified:   src/integrators/sppm.cc
	modified:   src/yafraycore/kdtree.cc
	modified:   src/yafraycore/mcintegrator.cc
	modified:   src/yafraycore/ray_kdtree.cc
	modified:   src/yafraycore/scene.cc