New Extended Texture Mapping System
Added many new texture mapping options for all materials, including filter color and IOR for glass and rough glass, sigma oren nayar for glossy and coated glossy, mirror amount and mirror color for coated glossy, etc.

This is very experimental, but I hope it helps the users to create new exotic materials and to reduce the usage of the blend material in some cases.

 Changes to be committed:
	modified:   include/materials/shinydiff.h
	modified:   src/materials/coatedglossy.cc
	modified:   src/materials/glossy.cc
	modified:   src/materials/shinydiffuse.cc