Removed unnecessary code from Glossy material and copied changes to glossy2
NOTE: I don't know why we keep glossy.cc and glossy2.cc if the two files are identical. I had a very strange problem because I made changes only to glossy.cc and it worked in Linux builds, but the changes did not appear in Windows builds. Unless some decision is made, KEEP THE TWO FILES WITH THE SAME CONTENTS to avoid that kind of discrepancies.

 Changes to be committed:
	modified:   src/materials/glossy.cc
	modified:   src/materials/glossy2.cc