Fix for Final gather using background even if no IBL/caustic enabled
In Photon Mapping and Path Tracing integrators, Final Gather was gathering the background even when IBL nor caustic were enabled in the background. This introduces undesired noise to the image from the background.

I've made changes to make sure Photon Mapping and Path Tracing *only* use the background for caustics calculations when the background has IBL *and* caustics enabled.

In the backgrounds like single color, gradient or SunSky1 where there is no parameter to enable caustic photons, I've set Final Gather to use the background just when IBL is enabled.

 Changes to be committed:
	modified:   include/core_api/background.h
	modified:   src/backgrounds/darksky.cc
	modified:   src/backgrounds/gradientback.cc
	modified:   src/backgrounds/sunsky.cc
	modified:   src/backgrounds/textureback.cc
	modified:   src/integrators/pathtracer.cc
	modified:   src/integrators/photonintegr.cc