Reducing white dots in Rough Glass material
The Rough Glass material is sometimes noisy showing white dots (fireflies) with very high intensity that cannot be smoothed out using AA.

I'm not sure about what exactly is the root cause for this problem, but I suspect the pdf calculation is somehow incorrect somewhere, but not sure why. This causes sometimes during the recursive raytracing stage to amplify the color values of certain surface points up to high values causing the white dots.

What I've tried to do to minimize/alleviate this situation is to increase the precission of the trig functions in the material and also to put a lower limit to the pdf (0.1 for now) so the amplification of the color values is not so big. This will cause the rough glass material to be slightly darker than before, but hopefully less and more manageable white dots noise.

For more information, examples and images see: http://yafaray.org/node/663

Changes to be committed:
modified: src/materials/roughglass.cc