Temporary workaround for artifacts in bump mapping
Dirty workaround/hack to avoid artifacts (slightly shadowed triangles with slightly different normals?? coincidently with changes in dSdU, dSdV, etc) when using image textures for bump mapping. This change effectively disables this entire calculation intended to be done for image textures, which is not really a good solution. We sould review why is this problem happening in the first place. Looks like some strange issue when using dSdU/dSdV for calculations???

 Changes to be committed:
	modified:   src/textures/basicnodes.cc