Changes to bump mapping/normal mapping
Bump mapping is very unstable and has several types of artifacs, so the render results are quite different from Blender Internal, for example. For more information about the problems see: http://www.yafaray.org/node/660

Problems found in current bump mapping:
* If the object does not have UV map associated, when using an image texture with mapping Global-Cube for example, more or less subtle triangular artifacts appear.

* If the object has UV map associated, when using an image texture with mapping Global-Cube for example, the image looks normal but if you look more carefully and you compare with an equivalent Blender Internal render, the normals are not correct, something is clearly off even if it's not obvious at the first glance.

* If the object does not have UV map associated, when using an image texture with mapping UV-Flat for example, the image is wrong (big triangles appear) and even worse, sometimes (and totally randomly) black dots (+NaN) appear. If you render a few times they might not appear, but if you render more times (aprox.10 times in a row) then you start seeing the black dots, and sometimes huge blocks of black NaN dots appear!!

The changes contained in this commit try to fix or at least minimize these issues by doing several changes to the bump mapping (that also affect normal mapping):
* The problem with +NaN black dots was found in a certain case where uninitialized u,v values were used. I forced u=v=0 in that case, so the surface will not show bump but no black dots will appear (similar to the result from Blender Internal)
* Depending if the surface point belongs to a mesh with UV map or not, and also depending if the texture mapping is "UV" or not, I apply different ways of generating the bump/normal differentials, to try to mimic the same results obtained in Blender Internal and get a more "realistic" look in the bump mapping.

However, due to the intrinsic nature of YafaRay and especially when no UV map is used and it has to be intrinsically generated, as we work with triangles it's difficult to generate automatically good UV values. Therefore all these fixes could be considered a series of tweaks and "hacks" but there is still possibility of having bump mapping artifacts in certain scenes. We need this change to be thoroughly tested but I hope that it improves the results over the original bump mapping that had so many issues.

 Changes to be committed:
	modified:   src/textures/basicnodes.cc
	modified:   src/yafraycore/triangle.cc