New parameter - cast shadows
As requested in http://www.yafaray.org/node/665 a new Material parameter has been added to enable/disable casting shadows. The default value is to be enabled, as that's the expected behavior for all objects in a scene. However, if certain objects need to be created shadowless (like in some architectural images), then shadow casting can be disabled in a per-material basis.

 Changes to be committed:
	modified:   include/core_api/material.h
	modified:   include/materials/blendmat.h
	modified:   include/materials/maskmat.h
	modified:   include/materials/roughglass.h
	modified:   include/materials/shinydiff.h
	modified:   src/materials/blend.cc
	modified:   src/materials/coatedglossy.cc
	modified:   src/materials/glass.cc
	modified:   src/materials/glossy.cc
	modified:   src/materials/glossy2.cc
	modified:   src/materials/mask.cc
	modified:   src/materials/roughglass.cc
	modified:   src/materials/shinydiffuse.cc
	modified:   src/yafraycore/kdtree.cc