Fix for error message "Index out of bounds in pdf1D_t"
Ocasionally we get the message "Index out of bounds in pdf1D_t::Sample: index, u, ptr, cdf = ". This is an old problem that used to cause black dots and was "patched" with this message and forcing the index to be within bounds, but the actual cause for the index going out of bounds was never found.

povmaniaco has reported the problem in the bug tracker and provided a way to replicate it with spotlights with falloff > 0.70 and caustic photons enabled. More information:
http://yafaray.org/node/681

I could replicate it with a single spotlight and this problem also affects background lights with, for example, EXR textures + IBL + use caustics

Now, I believe the problem is caused by scrHalton function, as it sometimes gives a result exactly=0.f. When in mcintegrator we get s2=0.f exactly, if the light uses Sample(s2, ...) for photon emission, the Sample function has ptr=cdf and therefore the index becomes -1, out of bounds.

So, I've added a minimum value to the scrHalton function. I tried to use a very very small minimum value (it was 0.f before, and I changed it to 1.0e-36, which is really close to 0.f). It seems to fix the problem and I think such a small minimum value should not cause other problems in the render process (I hope!)

 Changes to be committed:
	modified:   include/utilities/sample_utils.h
	modified:   include/yafraycore/scr_halton.h