Fix for attenuation when using 2 or more volumetric objects
Reported in bug tracker:
http://yafaray.org/node/332

This is a very old problem. I'm not 100% sure that what I've done is correct, but in my opinion having different attenuation in the volumetric objects depending on how many volumes you have do not seem right to me.

The problem is that, in the SingleScatterIntegrator, for some reason, the attenuation is calculated depending on the number of Volume Regions in the scene. I don't understand why was it done that way, maybe there is a reason for it??

In any case, to fix the problem mentioned in the bugtracker and having the same attenuation independently of the number of volumes in the scene, I've modified the code as you can see in this commit.

Probably a deeper review is needed before this is merged into YafaRay master.

 Changes to be committed:
	modified:   src/integrators/SingleScatterIntegrator.cc