Disable by default transparent background and transp.refracted background
To properly render volumes with a transparent background, in the previous changes to Volumetric rendering, when enabling "transparent refracted background", in addition to making the background transparent it's not even rendered at all, so the volume rendering does not contain "remains" of the background.

However, this new behavior of "transparent refracted background" can cause unexpeced results when rendering older XML files or XML files that lack those parameters. If those parameters are missing, the default value for transparent background and transparent refracted background is true, so the background would not be rendered at all. If in yafaray-xml you don't use the "-a" option to generate tranparency, you would get a black background, etc.

So, now I'm setting both transparent background and transparent refracted background default values to "false", to avoid those issues. Of course they can be overrided and set either to false or true just by adding them to the XML file.

 Changes to be committed:
	modified:   src/integrators/bidirpath.cc
	modified:   src/integrators/directlight.cc
	modified:   src/integrators/pathtracer.cc
	modified:   src/integrators/photonintegr.cc
	modified:   src/integrators/sppm.cc