Material visibility - code simplification
No changes in functionality here, just simplifying the code for material visibility so member variables and functions are defined in the master material.h file and not in each material (as visibility is a common property for all materials). Also, as getVisibility is doing the same in all derived classes, I've tried to improve a bit performance by replacing the virtual function by a normal function, hopefully it should make this a bit faster.

 Changes to be committed:
	modified:   include/core_api/material.h
	modified:   include/materials/blendmat.h
	modified:   include/materials/maskmat.h
	modified:   include/materials/roughglass.h
	modified:   include/materials/shinydiff.h
	modified:   src/materials/blend.cc
	modified:   src/materials/coatedglossy.cc
	modified:   src/materials/glass.cc
	modified:   src/materials/glossy.cc
	modified:   src/materials/glossy2.cc
	modified:   src/materials/mask.cc
	modified:   src/materials/roughglass.cc
	modified:   src/materials/shinydiffuse.cc