Automatic AA Threshold reduction based on a "floor" for resampled pixels
For better noise reduction, as requested in the bugtracker http://yafaray.org/node/690 I've added the functionality of automatically decreasing the AA Threshold before the next pass if the amount of resampled pixels is below a certain "floor" value selectable by the user.

This first version of the feature will have a hardcoded value of 0.9f for AA thresold reduction. This means that, every time a pass resamples fewer pixels than the "floor" selected by the user, the AA Threshold will be decreased by 10% before the next pass.

 Changes to be committed:
	modified:   include/core_api/imagefilm.h
	modified:   include/core_api/scene.h
	modified:   include/core_api/tiledintegrator.h
	modified:   src/yafraycore/environment.cc
	modified:   src/yafraycore/imagefilm.cc
	modified:   src/yafraycore/integrator.cc
	modified:   src/yafraycore/scene.cc