Disable differentials for now
Differential rays are implemented but not used for now. They are expected to be used in future features like motion blur or interference, but for now the only they are doing is to waste CPU processing power. So I've added a new boolean parameter to disable them for now until those features are implemented.

 Changes to be committed:
	modified:   include/core_api/tiledintegrator.h
	modified:   src/integrators/photonintegr.cc
	modified:   src/integrators/sppm.cc
	modified:   src/yafraycore/integrator.cc
	modified:   src/yafraycore/mcintegrator.cc