Fix for black dots using HDRI spherical and Coated Glossy
Black dots reported in http://www.yafaray.org/community/forum/viewtopic.php?p=31358#p31358 and http://www.yafaray.org/node/705 when using HDRI spherical and Coated Glossy material

I've found that the wi vector was sometimes not defined by the Coated Glossy sample function and it caused the next recursive step to use refRay=(0,0,0) which caused a problem in the texture spheremap function to convert 0,0,0 to u,v coordinates (division by zero)

I'm not sure why this happens sometimes in the Windows 64 build but not in the Linux 64 build (at least it did not happen to me) or why it happens in v2.0.1 but not in v1.1.0... strange but when dealing with not initialized values anything can happen.

So, for now I'll just add a wi assignment in case the sample function ends prematurely, I hope it solves this problem and does not cause any new strange issues.