Alternative Fix for black dots (alpha=0) sometimes when using Mitchell/Lanczos filters
In my previous commit, I fixed this problem by clipping the negative lobe of these filters. That allows to avoid the problems due to negative pixels but perhaps it could cause other problems like lower ability to sharpen the image, loss of energy conservation, etc.

So, in this commit I'm removing the fix above and instead of clipping negative values from the filter, I'm trying to handle better those negative values in YafaRay.

Due to negative filter output, sometimes a pixel could arrive to the image film with negative color components and negative weight. Dividing the components by the weight would give a correct pixel color, but until now, YafaRay was only normalizing positive values and in case of negative weight it set the color to (0,0,0,0), that's the reason for these black dots (alpha=0).

So, in this change I'm enabling the filter negative lobe again but changing the way the normalization is handled to allow negative values. Only the 0.0 value itself would be excluded.

This is also fixing the problem described in http://yafaray.org/node/712 and the resulting image is more similar to the original image with the black dots than the image rendered with the previous fix in the previous commit.

However, this is an important change and perhaps it could cause new artifacts and issues. As far as I know there is not a "good" solution for this problem and we would have to choose one of these two possible solutions. For now I will use this one instead of the previous. Let's see what happens.