Fix for incorrect IBL lighting in World Background Texture when set to Angular
For some weird reason, there was code in YafaRay that reversed the direction of the rays when using Angular Texture for World Background, causing the IBL to illuminate the opposite side of the scene respect to the actual lights!!

I don't know why was that done that way, maybe there is a reason?  Anyway this has been probably wrong since 2010!  I think this is the commit that caused this problem: https://github.com/DavidBluecame/Core/commit/63f57e2f6e80697d93c3b416e64deeb46f5556f3

In any case this seems to be solving the problem. I hope it does not cause new issues, but what can be worse than having IBL in the wrong place?