Fix crash when using Textures with Windows or Normal coords and Caustic Photons
When using Windows or Normal coordinates for textures, which are view-dependant, the mcintegrator Caustic Photons generation did not have the camera assigned, causing a crash.

I've assigned a camera to the state before generating the Caustic Photons to avoid the crash. However, we have to consider that Photons are supposed to be view-independent, but when using textures with Window or Normal coordinates (which are view dependant), then the Photons WILL also be view dependant as well!

So, using textures with "tricks" like Normal or Windows coordinates can cause low frequency noise in animations with Photons.