Spotlight with soft shadows too dark when distance < 1.f
As reported in http://www.yafaray.org/node/587, when using spotlights with soft shadows, if the distance between the spotlight and the surface is less than 1.f units, the light gets darker and not brighter as expected.

I've investigated this for quite a while and I've found that the light PDF is proportional to the square of the distance and goes below 1.f when the distance is below 1.f. That causes the amount of light to actually decrease instead of increasing further when it gets closer and closer to the surface.

I don't quite understand yet how is the spotlight sampling and PDF calculated, something is very wrong somewhere. In fact, the amount of light when the distance is below 2.f is quite different when comparing the spotlight without soft shadows and the spotlight with soft shadows. Something is very wrong here.

In any case, I've made a change to prevent the PDF to become less than 1.f, "moving" the effect of the distance to the color itself when the distance is less than 1.f. This is a horrible hack, and it still does not solve the problem with the "soft" spotlight intensity being around half normal intensity respect to non-soft spotlight, but it should at least allow users to have a decent spotlight when distances are less than 1 unit.