Fix for SPPM sudden brightness change when reaching approx 4,300 million photons As reported in http://www.yafaray.org/node/772  when SPPM accumulates a total of 4,300 million photons approximately, the scene brightness changes very suddenly.
I found the cause in the sppm total photon count was using a 32bit unsigned integer, which overflows after it reaches 4,294,967,295 (2^32 - 1)

I've replaced the total photon count variable with a 64bit unsigned integer (uint64_t) and it seems to be working fine now.

Note: I've seen other total photon counters in the same file that are using int64_t. I think they should use uint64_t instead (we are wasting half of the 64bit integer range with a sign that we don't really need), but I don't want to make changes that are not really needed to fix this bug, as I don't want to introduce any new bugs. However, perhaps in the future all these accumulated counters should be harmonized to uint64_t