New texture functionality: distance color averaging
As described in http://www.yafaray.org/node/695 there is a limitation in the current YafaRay code respect to textures.

YafaRay samples the textures "as is", independently of the distance between the camera and the object. That causes far objects to show noise in the texture due to high-frequency sampling problems (texture aliasing).

Normally this gets mitigated by increasing the number of samples. However, that's a very inefficient way and takes a very long time. Sometimes it's not even possible to remove the noise.

The ideal solution would be to implement an automatic texture "mipmapping" system, where we used different texture resolutions for sampling based on ray differentials. We must aim for that solution, but it will take quite some time.

Therefore, for a short term "quick fix" I've implemented a new functionality that allows to use the original texture for objects near the camera and just the average color of the texture for far objects. In the middle, a "blend" will be progressively done between the original texture and the averaged color.

This is FAR from a good solution but it might solve some of the current noise problems.

The way this works is:
* Parameter to enable/disable this function.
* Parameter to set the minimum distance. Any surfaces nearer than this distance will show the original texture
* Parameter to set the maximum distance. Surfaces farther away than the maximum distance will only show the averaged color. Between min and max, a progressive "blend" will be done between the original texture and the averaged color