IMPORTANT: big changes to textures interpolation and colorspace and BIG changes (again) to imagehandlers
To solve the problem detected at http://yafaray.org/node/787 where I found out that YafaRay is doing the texture interpolation *after* decoding the texels color space. This was causing significant differences in color between standard bilinear/bicubic and when using trilinear or EWA mipmaps.

I believe that the correct way to interpolate textures is to decode color space into linear space every time a texel is read, before the texels are used for interpolation. That way the interpolation is calculated correctly in linear color space.

Unfortunately there are negative consequences of decoding color space before interpolation: it will take many more calculations as color space is decoded, not on the result of the interpolation, but for every time a texel is read to be interpolated. In bilinear that means to multiply the color space decoding calculations by 4. In bicubic by 16!!  In trilinear by 8 and in EWA... it depends on each case but it could be a lot!!

In order to avoid that, and to provide a background compatibility path, I've created a new per-texture setting:
* ColorSpace Interpolation parameter in Blender-Exporter. This is both per-texture or if in the texture is set to "Default" then it will use the global setting with the same name in the main render tab
* <colorspace_interpolation_method sval="xxxxxx"/> in XML interface

The way this will work is as follows:
* <colorspace_interpolation_method sval="old-legacy"/>  This will use the original YafaRay color space decoding after interpolation. However, it's not recommended anymore because it's not mathematically correct and it could show differences when using mipmaps respect to non-mipmap interpolation
* <colorspace_interpolation_method sval="linear-memory"/>  This will use the new (hopefully mathematically correct!) interpolation after color space decoding. However, rendering *will* be slower than the old system, as many more color space decoding calculations need to take place.
* <colorspace_interpolation_method sval="linear-speed"/>  This will use the new (hopefully mathematically correct!) interpolation after color space decoding. To avoid the speed penalty of the "linear-memory" method, in this case I created a "clone" of the original texture in RAM with the colors already decoded to linear. Therefore the interpolation will get values from linear texels in RAM and not having to do any color space decoding during render. This will make rendering speed faster, probably even a bit faster than the original old-legacy YafaRay method. However, this comes with a penalty in RAM usage, as the memory to store textures will double!!

I also had another problem while implemented the new methods explained above. The linear-speed method, which is based on generating linear color copies of the original textures, will make the copies in the same internal Color Buffer format as the original. If the texture is using Optimization=optimized, until now, YafaRay stored the original texture in 8-bit per color channel. However, when creating linear copies of the original textures, 8-bit per channel will not have enough bit depth to store all the details of the decoded color. Therefore I had to increase the "optimized" color buffer size from RGB888 (3 bytes / 8 bit per channel) to RGB101010 (4 bytes / 10 bit per channel) and the RGBA8888 (4 bytes / 8 bits per channel) to RGBA1010108 (5 bytes / 10 bits per channel for RGB and 8 bits for A). This way I hoped to be able to keep all color information even after decoding the color space.

Therefore, when using "optimized" textures, the RAM usage will increase in around 20-25% respect to the previous v3.1.1 version. I could not find any reasonable way to avoid it while keeping the color accuracy. If RAM is a concern for a certain render, the user can choose "linear-memory" which is slower but uses a lot less RAM, and also the user can choose the optimization="compressed" which reduces the RAM usage further at the cost of losing some color precision/information.

Finally, I've used the opportunity to simplify and better organize the code and API for imagehandlers and textures, putting together all common code into the base class and simplifying the getColor and getFloat interface.

 Changes to be committed:
	modified:   include/core_api/imagehandler.h
	modified:   include/core_api/texture.h
	modified:   include/textures/basictex.h
	modified:   include/textures/imagetex.h
	modified:   include/utilities/image_buffers.h
	modified:   src/backgrounds/textureback.cc
	modified:   src/bindings/yafaray_v3_interface.i
	modified:   src/image_handlers/exrHandler.cc
	modified:   src/image_handlers/hdrHandler.cc
	modified:   src/image_handlers/jpgHandler.cc
	modified:   src/image_handlers/pngHandler.cc
	modified:   src/image_handlers/tgaHandler.cc
	modified:   src/image_handlers/tifHandler.cc
	modified:   src/textures/basicnodes.cc
	modified:   src/textures/basictex.cc
	modified:   src/textures/imagetex.cc
	modified:   src/yafraycore/imagefilm.cc
	modified:   src/yafraycore/imagehandler.cc